backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
appnope==0.1.4
asttokens==3.0.1
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
comm==0.2.3
debugpy==1.8.20
decorator==5.2.1
distro==1.9.0
executing==2.2.1
fastapi==0.136.0
groq==0.37.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipykernel==7.2.0
ipython==9.12.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.14.0
jsonpatch==1.33
jsonpointer==3.1.1
jupyter_client==8.8.0
jupyter_core==5.9.1
langchain==1.2.15
langchain-core==1.2.28
langchain-groq==1.1.2
langchain-openai==1.1.12
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.13
langsmith==0.7.30
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
openai==2.31.0
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.0
parso==0.8.6
pexpect==4.9.0
platformdirs==4.9.6
prompt_toolkit==3.0.52
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.13.0
pydantic_core==2.46.0
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
pyzmq==27.1.0
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==1.0.0
tavily-python==0.7.23
tenacity==9.1.4
tiktoken==0.12.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.44.0
wcwidth==0.6.0
xxhash==3.6.0
zstandard==0.25.0
